+++ src/react-query/queryConversations.ts
@@ -12,6 +12,17 @@ import { queryConversations } from "../funcs/queryConversations.js";
 import { combineSignals } from "../lib/primitives.js";
 import { RequestOptions } from "../lib/sdks.js";
 import * as components from "../models/components/index.js";
+import {
+  ConnectionError,
+  InvalidRequestError,
+  RequestAbortedError,
+  RequestTimeoutError,
+  UnexpectedClientError,
+} from "../models/errors/httpclienterrors.js";
+import * as errors from "../models/errors/index.js";
+import { InkeepAnalyticsError } from "../models/errors/inkeepanalyticserror.js";
+import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
+import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import { unwrapAsync } from "../types/fp.js";
 import { useInkeepAnalyticsContext } from "./_context.js";
 import { MutationHookOptions } from "./_types.js";
@@ -24,18 +35,33 @@ export type QueryConversationsMutationVariables = {
 export type QueryConversationsMutationData =
   components.QueryConversationsResponseBody;
 
+export type QueryConversationsMutationError =
+  | errors.BadRequest
+  | errors.Unauthorized
+  | errors.Forbidden
+  | errors.UnprocessableEntity
+  | errors.InternalServerError
+  | InkeepAnalyticsError
+  | ResponseValidationError
+  | ConnectionError
+  | RequestAbortedError
+  | RequestTimeoutError
+  | InvalidRequestError
+  | UnexpectedClientError
+  | SDKValidationError;
+
 /**
  * Query Conversations
  */
 export function useQueryConversationsMutation(
   options?: MutationHookOptions<
     QueryConversationsMutationData,
-    Error,
+    QueryConversationsMutationError,
     QueryConversationsMutationVariables
   >,
 ): UseMutationResult<
   QueryConversationsMutationData,
-  Error,
+  QueryConversationsMutationError,
   QueryConversationsMutationVariables
 > {
   const client = useInkeepAnalyticsContext();